feat: arbitrumm chain coverage improved

+++ src/integrations/chainlink/IFeedRegistryInterface.sol
@@ -7,6 +7,5 @@ interface IFeedRegistryInterface {
     function latestRoundData(
         address base,
         address quote
-    ) external view returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
-
+    ) external view returns (uint80 roundId, int answer, uint startedAt, uint updatedAt, uint80 answeredInRound);
 }
